Handle pending subscriptions by creating a PUT request
**Situation:**
Double opt-in which requires an e-mail verification. User doesn't get
the e-mail (spam filter, temporary error etc.) and never confirms.
Later the user retries the signup process, but MailChimp will report
"already signed up" even though the signup is pending.

**Solution:**
If the user is not already fully subscribed, send a PUT request to
MailChimp. This will trigger the signup mail again and solve the signup
limbo.

If the user is already fully subscribed nothing, the existing
subscription is not touched and the existing "error message" is shown.